+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- Experimental support for token-based authentication in the client and functions for token management.
+
 ## [8.15.0]
 
 ### Added